Use a new sector-classification bridge in r2dii.data

Use a new sector-classification bridge in r2dii.data

use_bridge(dataset, data, contributor = NULL, issue = NULL, overwrite = FALSE)

Arguments

dataset
String. Name of the dataset, to be used as the prefix of an object with the format [prefix]_classification, e.g. psic_classification.

data
A data frame.

contributor
String. Name of a contributor to thank in NEWS.md, e.g. "\@daisy-pacheco".

issue
String. Number of related issue or PR, e.g. #199.

overwrite
Logical. Allow overwriting data?

Value

Most functions are interactive, and called for their side effects. The return value is usually invisible(dataset).

Examples

if (FALSE) {
# Setup -------------------------------------------------------------------
# Work form inside r2dii.data
fs::dir_copy("../r2dii.data", tempdir())
tmp <- fs::path(tempdir(), "r2dii.data")
old <- getwd()
setwd(tmp)

# Call use_bridge ---------------------------------------------------------
library(r2dii.usethis)

# This is usually a contributed spreadsheet
dataset <- "fake"
data <- data.frame(
  original_code = letters[1:3],
  code_level = 5,
  code = as.character(1:3),
  sector = "not in scope",
  borderline = FALSE
)
contributor <- "@somebody"
issue <- "#123"
use_bridge(dataset, data, contributor, issue)

# Notice what changed
system("git status -s")

# Teardown ----------------------------------------------------------------
unlink(tmp, recursive = TRUE)
setwd(old)
}