A checklist to add new data to the package r2dii.data

A checklist to add new data to the package r2dii.data

use_r2dii_data(name = NULL)

Arguments

name

String of length 1 giving the name of the dataset, e.g. "ald_demo".

Value

Printed output that you may copy and paste into a GitHub issue.

Examples

# Maybe copy and paste into a GitHub issue.
# See `?usethis::browse_github_issues()`
use_r2dii_data()
#> Add `newdata`
#> 
#> * [ ] Add "data-raw/newdata.csv" 
#> * [ ] Add "data-raw/data_dictionary/newdata.csv" 
#> * [ ] In "data-raw/newdata.R", add something like: 
#>   ```
#>   library(usethis)
#>   # Source: @<contributor> <URL to issue or pull request>
#>   newdata <- read_csv_(file.path("data-raw", "newdata.csv"))
#>   use_data(newdata, overwrite = TRUE)
#>   ```
#> * [ ] `usethis::use_r("newdata")`. Document `newdata` 
#> * [ ] `usethis::use_test("newdata")`. Add regression tests of `newdata` 
#> * [ ] `source("data-raw/data_dictionary.R"); devtools::load_all()` 
#> * [ ] `usethis::use_test("data_dictionary")`. Test `data_dictionary` includes `newdata` 
#> * [ ] Source data-raw/ and test all datasets remain the same: 
#>   ```
#>   r2dii.data::source_data_raw()
#>   devtools::load_all()
#>   devtools::test()
#>   devtools::check()
#>   ```
#> * [ ] Edit NEWS.md to include newdata. 
#> * [ ] Increment version number in DESCRIPTION and NEWS.md: `usethis::use_version()` 

use_r2dii_data("my_data")
#> Add `my_data`
#> 
#> * [ ] Add "data-raw/my_data.csv" 
#> * [ ] Add "data-raw/data_dictionary/my_data.csv" 
#> * [ ] In "data-raw/my_data.R", add something like: 
#>   ```
#>   library(usethis)
#>   # Source: @<contributor> <URL to issue or pull request>
#>   my_data <- read_csv_(file.path("data-raw", "my_data.csv"))
#>   use_data(my_data, overwrite = TRUE)
#>   ```
#> * [ ] `usethis::use_r("my_data")`. Document `my_data` 
#> * [ ] `usethis::use_test("my_data")`. Add regression tests of `my_data` 
#> * [ ] `source("data-raw/data_dictionary.R"); devtools::load_all()` 
#> * [ ] `usethis::use_test("data_dictionary")`. Test `data_dictionary` includes `my_data` 
#> * [ ] Source data-raw/ and test all datasets remain the same: 
#>   ```
#>   r2dii.data::source_data_raw()
#>   devtools::load_all()
#>   devtools::test()
#>   devtools::check()
#>   ```
#> * [ ] Edit NEWS.md to include my_data. 
#> * [ ] Increment version number in DESCRIPTION and NEWS.md: `usethis::use_version()`